Move command and event to client module
- fix remove connection when client connection away
