+++ Cargo.toml
@@ -25,7 +25,6 @@ slog-stdlog = "3.0"
 slog-async = "2.3"
 slog-json = "2.3"
 slog-scope = "4.1"
-failure = "0.1"
 
 [dev-dependencies]
 tokio = "0.1"
