+++ src/api/client/commands.rs
@@ -1,0 +1,25 @@
+use actix::Message;
+use serde::{Deserialize, Serialize};
+
+use crate::{api::client::RoomError, media::peer::Id as PeerId};
+
+/// WebSocket message from Web Client to Media Server.
+#[derive(Debug, Deserialize, Message, Serialize)]
+#[rtype(result = "Result<(), RoomError>")]
+pub enum Command {
+    /// Web Client sends SDP Offer.
+    MakeSdpOffer {
+        peer_id: PeerId,
+        sdp_offer: String,
+    },
+    /// Web Client sends SDP Answer.
+    MakeSdpAnswer {
+        peer_id: PeerId,
+        sdp_answer: String,
+    },
+
+    SetIceCandidate {
+        peer_id: PeerId,
+        candidate: String,
+    },
+}
