+++ src/api/client/mod.rs
@@ -1,6 +1,11 @@
 //! Implementation of Client API.
 
+pub mod commands;
+pub mod events;
+pub mod room;
 pub mod server;
 pub mod session;
 
-pub use self::{server::*, session::*};
+pub use self::{
+    commands::Command, events::Event, room::*, server::*, session::*,
+};
