+++ src/api/client/server.rs
@@ -108,11 +108,7 @@ mod test {
             1 => Member{id: 1, credentials: "caller_credentials".into()},
             2 => Member{id: 2, credentials: "responder_credentials".into()},
         };
-        let room = Arbiter::start(move |_| Room {
-            id: 1,
-            members,
-            connections: HashMap::new(),
-        });
+        let room = Arbiter::start(move |_| Room::new(1, members));
         let rooms = hashmap! {1 => room};
         RoomsRepository::new(rooms)
     }
