+++ src/media/mod.rs
@@ -1,15 +1,11 @@
 pub mod errors;
 pub mod peer;
-pub mod room;
 pub mod track;
 
 pub use self::{
     errors::MediaError,
-    peer::{Event, Id as PeerID, Peer, PeerMachine},
-    room::{
-        Command, Id as RoomID, Room, RoomsRepository, RpcConnection,
-        RpcConnectionClosed, RpcConnectionClosedReason,
-        RpcConnectionEstablished,
+    peer::{Id as PeerId, Peer, PeerMachine},
+    track::{
+        AudioSettings, Id as TrackId, Track, TrackMediaType, VideoSettings,
     },
-    track::Id as TrackID,
 };
